+++ src/c_api/c_api_common.h
@@ -1,0 +1,79 @@
+/*!
+ *  Copyright (c) 2015 by Contributors
+ * \file c_api_error.h
+ * \brief Error handling for C API.
+ */
+#ifndef MXNET_C_API_C_API_ERROR_H_
+#define MXNET_C_API_C_API_ERROR_H_
+
+#include <dmlc/base.h>
+#include <dmlc/logging.h>
+#include <dmlc/thread_local.h>
+#include <mxnet/c_api.h>
+#include <mxnet/base.h>
+
+/*! \brief  macro to guard beginning and end section of all functions */
+#define API_BEGIN() try {
+/*! \brief every function starts with API_BEGIN();
+     and finishes with API_END() or API_END_HANDLE_ERROR */
+#define API_END() } catch(dmlc::Error &_except_) { return MXAPIHandleException(_except_); } return 0;  // NOLINT(*)
+/*!
+ * \brief every function starts with API_BEGIN();
+ *   and finishes with API_END() or API_END_HANDLE_ERROR
+ *   The finally clause contains procedure to cleanup states when an error happens.
+ */
+#define API_END_HANDLE_ERROR(Finalize) } catch(dmlc::Error &_except_) { Finalize; return MXAPIHandleException(_except_); } return 0; // NOLINT(*)
+
+/*!
+ * \brief Set the last error message needed by C API
+ * \param msg The error message to set.
+ */
+void MXAPISetLastError(const char* msg);
+/*!
+ * \brief handle exception throwed out
+ * \param e the exception
+ * \return the return value of API after exception is handled
+ */
+inline int MXAPIHandleException(const dmlc::Error &e) {
+  MXAPISetLastError(e.what());
+  return -1;
+}
+
+using namespace mxnet;
+
+/*! \brief entry to to easily hold returning information */
+struct MXAPIThreadLocalEntry {
+  /*! \brief result holder for returning string */
+  std::string ret_str;
+  /*! \brief result holder for returning strings */
+  std::vector<std::string> ret_vec_str;
+  /*! \brief result holder for returning string pointers */
+  std::vector<const char *> ret_vec_charp;
+  /*! \brief result holder for returning handles */
+  std::vector<void *> ret_handles;
+  /*! \brief result holder for returning shapes */
+  std::vector<TShape> arg_shapes, out_shapes, aux_shapes;
+  /*! \brief result holder for returning type flags */
+  std::vector<int> arg_types, out_types, aux_types;
+  /*! \brief result holder for returning shape dimensions */
+  std::vector<mx_uint> arg_shape_ndim, out_shape_ndim, aux_shape_ndim;
+  /*! \brief result holder for returning shape pointer */
+  std::vector<const mx_uint*> arg_shape_data, out_shape_data, aux_shape_data;
+  // helper function to setup return value of shape array
+  inline static void SetupShapeArrayReturn(
+      const std::vector<TShape> &shapes,
+      std::vector<mx_uint> *ndim,
+      std::vector<const mx_uint*> *data) {
+    ndim->resize(shapes.size());
+    data->resize(shapes.size());
+    for (size_t i = 0; i < shapes.size(); ++i) {
+      ndim->at(i) = shapes[i].ndim();
+      data->at(i) = shapes[i].data();
+    }
+  }
+};
+
+// define the threadlocal store.
+typedef dmlc::ThreadLocalStore<MXAPIThreadLocalEntry> MXAPIThreadLocalStore;
+
+#endif  // MXNET_C_API_C_API_ERROR_H_
