+++ src/c_api/c_api_error.cc
@@ -3,14 +3,13 @@
  * \file c_api_error.cc
  * \brief C error handling
  */
-#include "./c_api_error.h"
-#include "../common/thread_local.h"
+#include "./c_api_common.h"
 
 struct ErrorEntry {
   std::string last_error;
 };
 
-typedef mxnet::common::ThreadLocalStore<ErrorEntry> MXAPIErrorStore;
+typedef dmlc::ThreadLocalStore<ErrorEntry> MXAPIErrorStore;
 
 const char *MXGetLastError() {
   return MXAPIErrorStore::Get()->last_error.c_str();
