--- src/c_api/c_api_error.h
@@ -1,39 +1,0 @@
-/*!
- *  Copyright (c) 2015 by Contributors
- * \file c_api_error.h
- * \brief Error handling for C API.
- */
-#ifndef MXNET_C_API_C_API_ERROR_H_
-#define MXNET_C_API_C_API_ERROR_H_
-
-#include <dmlc/base.h>
-#include <dmlc/logging.h>
-#include <mxnet/c_api.h>
-
-/*! \brief  macro to guard beginning and end section of all functions */
-#define API_BEGIN() try {
-/*! \brief every function starts with API_BEGIN();
-     and finishes with API_END() or API_END_HANDLE_ERROR */
-#define API_END() } catch(dmlc::Error &_except_) { return MXAPIHandleException(_except_); } return 0;  // NOLINT(*)
-/*!
- * \brief every function starts with API_BEGIN();
- *   and finishes with API_END() or API_END_HANDLE_ERROR
- *   The finally clause contains procedure to cleanup states when an error happens.
- */
-#define API_END_HANDLE_ERROR(Finalize) } catch(dmlc::Error &_except_) { Finalize; return MXAPIHandleException(_except_); } return 0; // NOLINT(*)
-
-/*!
- * \brief Set the last error message needed by C API
- * \param msg The error message to set.
- */
-void MXAPISetLastError(const char* msg);
-/*!
- * \brief handle exception throwed out
- * \param e the exception
- * \return the return value of API after exception is handled
- */
-inline int MXAPIHandleException(const dmlc::Error &e) {
-  MXAPISetLastError(e.what());
-  return -1;
-}
-#endif  // MXNET_C_API_C_API_ERROR_H_
